Add metadata to SyncConfig
`kube.NetworkPolicy` creates an object with empty `metadata.annotations: {}`.
This causes a problem when ArgoCD tries to create a SyncConfig spec with that,
but Espejo removes the empty Annotation field when updating its status.

Thus ArgoCD is always OutOfSync.

This can be fixed by either removing the field, or by adding metadata in form of annotations.