Add support for restricting the `allow-from-cluster-nodes` Cilium network policy
We introduce a new parameter `allowFromNodeLabels` which can be used to
customize the `allow-from-cluster-nodes` CiliumNetworkPolicy. By
default, the component deploys a permissive policy that allows access to
workloads from all cluster nodes (also across clusters in a cluster-mesh
setup).

By setting the parameter, and enabling Cilium's `nodeSelectorLabels`
policy feature (Cilium 1.16 and newer), it's possible to restrict access
to workloads to a subset of nodes in the cluster based on their labels.